+++ src/app/home/home.component.ts
@@ -34,6 +34,9 @@ import { SensorType } from '@interfaces/sensor-type';
 import { SensorTypeService } from '@app/services/sensor-type.service';
 import { DisplayValuePipe } from '@pipes/display-value.pipe';
 import { SitesService } from '@services/sites.service';
+import { SiteService } from '@services/site.service';
+import 'leaflet.markercluster';
+import 'leaflet.markercluster.freezable';
 import { FilteredEventsQuery } from '@interfaces/filtered-events-query';
 
 import { MatInputModule } from '@angular/material/input';
@@ -49,6 +52,7 @@ import * as esri from 'esri-leaflet';
 import { EventTypeService } from '@app/services/event-type.service';
 import { EventType } from '@app/interfaces/event-type';
 import { Subject } from 'rx';
+import { canvas } from 'leaflet';
 
 export interface PeriodicElement {
     name: string;
@@ -97,6 +101,9 @@ export class HomeComponent implements OnInit {
     siteClicked = false;
     siteName;
 
+    //for the ALl STN Sites layer
+    allSites: Site[];
+
     public selectedEvent;
     currentEvent: number; //change to subject?
     currentEventName: string;
@@ -142,15 +149,24 @@ export class HomeComponent implements OnInit {
     currWarningsVisible = false;
     watchWarnVisible = false;
     ahpsGagesVisible = false;
+    allSitesVisible = false;
 
     //Used for determining when to show layer visibility snack bar message
     currentZoom: number;
     previousZoom: number;
 
+    //for all map layers that aren't basemaps
+    supplementaryLayers;
+
     public mapFilterForm: FormGroup;
 
     private displayedSites: Subject<Site[]> = new Subject<Site[]>();
     private setDisplayedSites;
+
+    eventIcon = L.divIcon({
+        className:
+            ' wmm-pin wmm-altblue wmm-icon-circle wmm-icon-white wmm-size-20',
+    });
 
     // TODO:1) populate table of events using pagination. consider the difference between the map and the table.
     //      2) setup a better way to store the state of the data - NgRx.This ought to replace storing it in an object local to this component,
@@ -168,7 +184,8 @@ export class HomeComponent implements OnInit {
         public currentUserService: CurrentUserService,
         private sitesService: SitesService,
         private displayValuePipe: DisplayValuePipe,
-        public snackBar: MatSnackBar
+        public snackBar: MatSnackBar,
+        private siteService: SiteService
     ) {
         this.eventTypes$ = this.eventTypeService.eventTypes$;
 
@@ -253,6 +270,21 @@ export class HomeComponent implements OnInit {
         this.states$ = this.stateService.getStates();
         // create and configure map
         this.createMap();
+
+        const siteIcon = L.divIcon({
+            className: ' allSiteIcon ',
+            iconSize: 32,
+        });
+
+        //Add all the STN sites to a layer when the map loads
+        this.siteService.getAllSites().subscribe((results) => {
+            this.allSites = results;
+            this.mapResults(
+                this.allSites,
+                siteIcon,
+                this.siteService.siteMarkers
+            );
+        });
     }
 
     setCurrentFilter() {
@@ -308,7 +340,11 @@ export class HomeComponent implements OnInit {
             .getEventSites(this.currentEvent)
             .subscribe((results) => {
                 this.eventSites = results;
-                this.mapResults(this.eventSites);
+                this.mapResults(
+                    this.eventSites,
+                    this.eventIcon,
+                    this.eventMarkers
+                );
             });
     }
 
@@ -318,11 +354,22 @@ export class HomeComponent implements OnInit {
             center: MAP_CONSTANTS.defaultCenter,
             zoom: MAP_CONSTANTS.defaultZoom,
             layers: [MAP_CONSTANTS.mapLayers.tileLayers.osm],
+            renderer: L.canvas(),
         });
-        /* this.markers = L.featureGroup().addTo(this.map); */
+
+        this.supplementaryLayers = {
+            Watersheds: MAP_CONSTANTS.mapLayers.esriDynamicLayers.HUC,
+            'All STN Sites': this.siteService.siteMarkers,
+            'Current Warnings*':
+                MAP_CONSTANTS.mapLayers.esriFeatureLayers.currentWarnings,
+            'Watches/Warnings*':
+                MAP_CONSTANTS.mapLayers.esriFeatureLayers.watchesWarnings,
+            "<span>AHPS Gages*</span> <br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>":
+                MAP_CONSTANTS.mapLayers.esriFeatureLayers.AHPSGages,
+        };
 
         L.control
-            .layers(MAP_CONSTANTS.baseMaps, MAP_CONSTANTS.supplementaryLayers, {
+            .layers(MAP_CONSTANTS.baseMaps, this.supplementaryLayers, {
                 position: 'topleft',
             })
             .addTo(this.map);
@@ -385,11 +432,14 @@ export class HomeComponent implements OnInit {
             }
         });
 
-        // When the watershed checkbox is checked, add watershed icon to legend
+        // When layer is checked, add layer icon to legend
         /* istanbul ignore next */
         this.map.on('overlayadd', (e) => {
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = true;
+            }
+            if (e.name === 'All STN Sites') {
+                this.allSitesVisible = true;
             }
             if (e.name === 'Current Warnings*') {
                 this.currWarningsVisible = true;
@@ -399,16 +449,20 @@ export class HomeComponent implements OnInit {
             }
             if (
                 e.name ===
-                "<span>AHPS Gages*</span><br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>"
+                "<span>AHPS Gages*</span> <br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>"
             ) {
                 this.ahpsGagesVisible = true;
             }
         });
-        // When the watershed checkbox is unchecked, remove watershed icon from legend
+        // When layer is unchecked, remove layer icon from legend
         /* istanbul ignore next */
         this.map.on('overlayremove', (e) => {
+            console.log('this is e remove', e);
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = false;
+            }
+            if (e.name === 'All STN Sites') {
+                this.allSitesVisible = false;
             }
             if (e.name === 'Current Warnings*') {
                 this.currWarningsVisible = false;
@@ -418,7 +472,7 @@ export class HomeComponent implements OnInit {
             }
             if (
                 e.name ===
-                "<span>AHPS Gages*</span><br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>"
+                "<span>AHPS Gages*</span> <br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>"
             ) {
                 this.ahpsGagesVisible = false;
             }
@@ -431,6 +485,13 @@ export class HomeComponent implements OnInit {
         //Get the value of the current zoom
         this.map.on('zoomend', () => {
             this.currentZoom = this.map.getZoom();
+            //Disable clustering for the All STN Sites layer when zoom >= 12 so we can see individual sites
+            if (this.currentZoom >= 12) {
+                this.siteService.siteMarkers.disableClustering();
+            }
+            if (this.currentZoom < 12) {
+                this.siteService.siteMarkers.enableClustering();
+            }
             //If the zoom went from 9 to 8 and the gages/watches/warnings are on,
             //that layer is checked, but it's not displayed
             //warn users of that in a snack bar message
@@ -582,35 +643,48 @@ export class HomeComponent implements OnInit {
             });
         });
 
-        // When the watershed checkbox is checked, add watershed icon to legend
+        // When layer is checked, add layer icon to legend
         /* istanbul ignore next */
         this.map.on('overlayadd', (e) => {
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = true;
             }
-            if (e.name === 'Current Warnings') {
+            if (e.name === 'All STN Sites') {
+                this.allSitesVisible = true;
+            }
+            if (e.name === 'Current Warnings*') {
                 this.currWarningsVisible = true;
             }
-            if (e.name === 'Watches/Warnings') {
+            if (e.name === 'Watches/Warnings*') {
                 this.watchWarnVisible = true;
             }
-            if (e.name === 'AHPS Gages') {
+            if (
+                e.name ===
+                "<span>AHPS Gages*</span> <br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>"
+            ) {
                 this.ahpsGagesVisible = true;
             }
         });
-        // When the watershed checkbox is unchecked, remove watershed icon from legend
+        // When layer is unchecked, remove layer icon from legend
         /* istanbul ignore next */
         this.map.on('overlayremove', (e) => {
+            console.log('this is e remove', e);
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = false;
             }
-            if (e.name === 'Current Warnings') {
+            if (e.name === 'All STN Sites') {
+                this.allSitesVisible = false;
+            }
+            if (e.name === 'Current Warnings*') {
                 this.currWarningsVisible = false;
             }
-            if (e.name === 'Watches/Warnings') {
+            if (e.name === 'Watches/Warnings*') {
                 this.watchWarnVisible = false;
             }
-            if (e.name === 'AHPS Gages') {
+            if (
+                e.name ===
+                "<span>AHPS Gages*</span> <br> <div class='leaflet-control-layers-separator'></div><span style='color: gray; text-align: center;'>*Zoom to level 9 to enable</span>"
+            ) {
                 this.ahpsGagesVisible = false;
             }
         });
@@ -632,24 +706,17 @@ export class HomeComponent implements OnInit {
     }
 
     /* istanbul ignore next */
-    mapResults(eventSites: any) {
+    mapResults(eventSites: any, myIcon: any, layerType: any) {
         // set/reset resultsMarker var to an empty array
         const markers = [];
         const iconClass = ' wmm-icon-diamond wmm-icon-white ';
         const riverConditions = [];
-        this.eventMarkers.removeFrom(this.map);
-        this.eventMarkers.clearLayers();
 
-        // loop through results responsefrom a search query
-        if (this.eventSites.length !== undefined) {
-            for (const site of this.eventSites) {
+        // loop through results response from a search query
+        if (eventSites.length !== undefined) {
+            for (const site of eventSites) {
                 const lat = Number(site.latitude_dd);
                 const long = Number(site.longitude_dd);
-
-                const myicon = L.divIcon({
-                    className:
-                        ' wmm-pin wmm-altblue wmm-icon-circle wmm-icon-white wmm-size-20',
-                });
 
                 /* let popupContent = '';
 
@@ -662,9 +729,28 @@ export class HomeComponent implements OnInit {
         const popup = L.popup()
           .setContent(popupContent); */
                 /* istanbul ignore next */
-                L.marker([lat, long], { icon: myicon }).addTo(
-                    this.eventMarkers
-                );
+                if (isNaN(lat) || isNaN(long)) {
+                    console.log(
+                        'Skipped site ' +
+                            site.site_no +
+                            ' in All STN Sites layer due to null lat/lng'
+                    );
+                } else {
+                    //put all the event markers in the same layer group
+                    if (layerType == this.eventMarkers) {
+                        L.marker([lat, long], { icon: myIcon }).addTo(
+                            layerType
+                        );
+                    }
+                    //Make circle markers for the All STN Sites layer
+                    if (layerType == this.siteService.siteMarkers) {
+                        L.marker([lat, long], {
+                            icon: myIcon,
+                            iconSize: 32,
+                        }).addTo(layerType);
+                    }
+                }
+
                 /* .bindPopup(popup)
           .on('click',
             (data) => {
@@ -684,7 +770,10 @@ export class HomeComponent implements OnInit {
             }); */
             }
 
-            //add event markers to map
+            // myMarkers.addTo(this.map);
+        }
+
+        if (layerType == this.eventMarkers) {
             this.eventMarkers.addTo(this.map);
         }
     }
@@ -745,7 +834,9 @@ export class HomeComponent implements OnInit {
             bracketTrue;
 
         this.sitesService.getFilteredSites(urlParamString).subscribe((res) => {
-            this.mapResults(res);
+            this.mapResults(res, this.eventIcon, this.eventMarkers);
         });
+
+        return urlParamString;
     }
 }